pyproject.toml
[tool.poetry]
name = "super-biz-agent"
version = "1.0.0"
description = "基于 LangChain + LangGraph + FastAPI 的智能问答与运维系统"
authors = ["envy"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pydantic = "^2.10.0"
pydantic-settings = "^2.7.0"
langchain = "^0.3.0"
langchain-community = "^0.3.0"
langchain-openai = "^0.3.0"
langchain-milvus = "^0.1.0"
langgraph = "^0.2.0"
dashscope = "^1.22.0"
httpx = "^0.28.0"
python-multipart = "^0.0.18"
aiofiles = "^24.1.0"
sse-starlette = "^2.2.0"
rank-bm25 = "^0.2.2"
pyyaml = "^6.0"
redis = "^5.2.0"
aiomysql = "^0.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
